+++ src/NotepadLayer.cpp
@@ -21,7 +21,7 @@ bool NotepadLayer::init(){
 	
 	// Menus
 
-	auto pageMenu = CCMenu::create();
+	auto levelName = CCMenu::create();
 	auto touchFixes = CCMenu::create();
 	auto numLabels = CCLayer::create();
 	auto layout = CCMenu::create();
@@ -33,13 +33,13 @@ bool NotepadLayer::init(){
 	touchFixes->setPosition(-64, -19.0f);
 	touchFixes->setID("touch-fixes");
 
-	m_mainLayer->addChild(pageMenu, 1);
-	pageMenu->setPosition(-64.5f, -22.9f);
-	pageMenu->setID("page-menu");
+	m_mainLayer->addChild(levelName, 1);
+	levelName->setPosition(-64.5f, -22.9f);
+	levelName->setID("level-name");
 
 	m_buttonMenu->setTouchPriority(-507);
 	touchFixes->setTouchPriority(-506);
-	pageMenu->setTouchPriority(-507);
+	levelName->setTouchPriority(-507);
 
 	// bg
 
@@ -61,12 +61,15 @@ bool NotepadLayer::init(){
 		png = "GJ_square05.png";
 	}
 	else if (themePresets == "Geode Theme Color 1"){
+		// @geode-ignore(unknown-resource)
 		png = "geode.loader/GE_square01.png";
 	}
 	else if (themePresets == "Geode Theme Color 2"){
+		// @geode-ignore(unknown-resource)
 		png = "geode.loader/GE_square02.png";
 	}
 	else if (themePresets == "Geode Theme Color 3"){
+		// @geode-ignore(unknown-resource)
 		png = "geode.loader/GE_square03.png";
 	}
 
@@ -187,10 +190,10 @@ bool NotepadLayer::init(){
 	numLabels->addChild(tenth);
 
 	// TextInputs
-
+		
 	m_input1 = TextInput::create(200, "", "bigFont.fnt");
 	m_input1->setPosition(153, 234);
-	m_input1->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab1, "")));
+	m_input1->setString(Mod::get()->getSavedValue<std::string>(m_inputTab1));
 	m_input1->setCommonFilter(CommonFilter::Any);
 	m_input1->setCallback([this](std::string const& str){
 
@@ -201,7 +204,7 @@ bool NotepadLayer::init(){
 
 	m_input2 = TextInput::create(200, "", "bigFont.fnt");
 	m_input2->setPosition(153, 194);
-	m_input2->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab2, "")));
+	m_input2->setString(Mod::get()->getSavedValue<std::string>(m_inputTab2));
 	m_input2->setCommonFilter(CommonFilter::Any);
 	m_input2->setCallback([this](std::string const& str){
 
@@ -212,7 +215,7 @@ bool NotepadLayer::init(){
 
 	m_input3 = TextInput::create(200, "", "bigFont.fnt");
 	m_input3->setPosition(153, 154);
-	m_input3->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab3, "")));
+	m_input3->setString(Mod::get()->getSavedValue<std::string>(m_inputTab3));
 	m_input3->setCommonFilter(CommonFilter::Any);
 	m_input3->setCallback([this](std::string const& str){
 
@@ -223,7 +226,7 @@ bool NotepadLayer::init(){
 
 	m_input4 = TextInput::create(200, "", "bigFont.fnt");
 	m_input4->setPosition(153, 114);
-	m_input4->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab4, "")));
+	m_input4->setString(Mod::get()->getSavedValue<std::string>(m_inputTab4));
 	m_input4->setCommonFilter(CommonFilter::Any);
 	m_input4->setCallback([this](std::string const& str){
 
@@ -234,7 +237,7 @@ bool NotepadLayer::init(){
 
 	m_input5 = TextInput::create(200, "", "bigFont.fnt");
 	m_input5->setPosition(153, 74);
-	m_input5->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab5, "")));
+	m_input5->setString(Mod::get()->getSavedValue<std::string>(m_inputTab5));
 	m_input5->setCommonFilter(CommonFilter::Any);
 	m_input5->setCallback([this](std::string const& str){
 
@@ -245,7 +248,7 @@ bool NotepadLayer::init(){
 
 	m_input6 = TextInput::create(200, "", "bigFont.fnt");
 	m_input6->setPosition(153, 34);
-	m_input6->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab6, "")));
+	m_input6->setString(Mod::get()->getSavedValue<std::string>(m_inputTab6));
 	m_input6->setCommonFilter(CommonFilter::Any);
 	m_input6->setCallback([this](std::string const& str){
 
@@ -256,7 +259,7 @@ bool NotepadLayer::init(){
 
 	m_input7 = TextInput::create(200, "", "bigFont.fnt");
 	m_input7->setPosition(153, -6);
-	m_input7->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab7, "")));
+	m_input7->setString(Mod::get()->getSavedValue<std::string>(m_inputTab7));
 	m_input7->setCommonFilter(CommonFilter::Any);
 	m_input7->setCallback([this](std::string const& str){
 
@@ -267,7 +270,7 @@ bool NotepadLayer::init(){
 
 	m_input8 = TextInput::create(200, "", "bigFont.fnt");
 	m_input8->setPosition(153, -46);
-	m_input8->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab8, "")));
+	m_input8->setString(Mod::get()->getSavedValue<std::string>(m_inputTab8));
 	m_input8->setCommonFilter(CommonFilter::Any);
 	m_input8->setCallback([this](std::string const& str){
 
@@ -278,7 +281,7 @@ bool NotepadLayer::init(){
 
 	m_input9 = TextInput::create(200, "", "bigFont.fnt");
 	m_input9->setPosition(153, -86);
-	m_input9->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab9, "")));
+	m_input9->setString(Mod::get()->getSavedValue<std::string>(m_inputTab9));
 	m_input9->setCommonFilter(CommonFilter::Any);
 	m_input9->setCallback([this](std::string const& str){
 
@@ -289,7 +292,7 @@ bool NotepadLayer::init(){
 
 	m_input10 = TextInput::create(200, "", "bigFont.fnt");
 	m_input10->setPosition(153, -126);
-	m_input10->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab10, "")));
+	m_input10->setString(Mod::get()->getSavedValue<std::string>(m_inputTab10));
 	m_input10->setCommonFilter(CommonFilter::Any);
 	m_input10->setCallback([this](std::string const& str){
 
@@ -302,8 +305,8 @@ bool NotepadLayer::init(){
 	// Priority
 
     m_inputPriority1 = TextInput::create(50, "", "bigFont.fnt");
-	m_inputPriority1->setPosition(395, 234);
-	m_inputPriority1->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString1, "")));
+	m_inputPriority1->setPosition(348, 234);
+	m_inputPriority1->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString1));
 	m_inputPriority1->setCommonFilter(CommonFilter::Any);
 	m_inputPriority1->setCallback([this](std::string const& str){
 
@@ -313,8 +316,8 @@ bool NotepadLayer::init(){
 	layout->addChild(m_inputPriority1);
 
    	m_inputPriority2 = TextInput::create(50, "", "bigFont.fnt");
-	m_inputPriority2->setPosition(395, 194);
-	m_inputPriority2->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString2, "")));
+	m_inputPriority2->setPosition(348, 194);
+	m_inputPriority2->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString2));
 	m_inputPriority2->setCommonFilter(CommonFilter::Any);
 	m_inputPriority2->setCallback([this](std::string const& str){
 
@@ -324,8 +327,8 @@ bool NotepadLayer::init(){
 	layout->addChild(m_inputPriority2);
 
    	m_inputPriority3 = TextInput::create(50, "", "bigFont.fnt");
-	m_inputPriority3->setPosition(395, 154);
-	m_inputPriority3->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString3, "")));
+	m_inputPriority3->setPosition(348, 154);
+	m_inputPriority3->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString3));
 	m_inputPriority3->setCommonFilter(CommonFilter::Any);
 	m_inputPriority3->setCallback([this](std::string const& str){
 
@@ -335,8 +338,8 @@ bool NotepadLayer::init(){
 	layout->addChild(m_inputPriority3);
 
    	m_inputPriority4 = TextInput::create(50, "", "bigFont.fnt");
-	m_inputPriority4->setPosition(395, 114);
-	m_inputPriority4->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString4, "")));
+	m_inputPriority4->setPosition(348, 114);
+	m_inputPriority4->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString4));
 	m_inputPriority4->setCommonFilter(CommonFilter::Any);
 	m_inputPriority4->setCallback([this](std::string const& str){
 
@@ -346,8 +349,8 @@ bool NotepadLayer::init(){
 	layout->addChild(m_inputPriority4);
 
     m_inputPriority5 = TextInput::create(50, "", "bigFont.fnt");
-	m_inputPriority5->setPosition(395, 74);
-	m_inputPriority5->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString5, "")));
+	m_inputPriority5->setPosition(348, 74);
+	m_inputPriority5->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString5));
 	m_inputPriority5->setCommonFilter(CommonFilter::Any);
 	m_inputPriority5->setCallback([this](std::string const& str){
 
@@ -357,8 +360,8 @@ bool NotepadLayer::init(){
 	layout->addChild(m_inputPriority5);
 
 	m_inputPriority6 = TextInput::create(50, "", "bigFont.fnt");
-	m_inputPriority6->setPosition(395, 34);
-	m_inputPriority6->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString6, "")));
+	m_inputPriority6->setPosition(348, 34);
+	m_inputPriority6->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString6));
 	m_inputPriority6->setCommonFilter(CommonFilter::Any);
 	m_inputPriority6->setCallback([this](std::string const& str){
 
@@ -368,8 +371,8 @@ bool NotepadLayer::init(){
 	layout->addChild(m_inputPriority6);
 
     m_inputPriority7 = TextInput::create(50, "", "bigFont.fnt");
-	m_inputPriority7->setPosition(395, -6);
-	m_inputPriority7->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString7, "")));
+	m_inputPriority7->setPosition(348, -6);
+	m_inputPriority7->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString7));
 	m_inputPriority7->setCommonFilter(CommonFilter::Any);
 	m_inputPriority7->setCallback([this](std::string const& str){
 
@@ -379,8 +382,8 @@ bool NotepadLayer::init(){
 	layout->addChild(m_inputPriority7);
 
    	m_inputPriority8 = TextInput::create(50, "", "bigFont.fnt");
-	m_inputPriority8->setPosition(395, -46);
-	m_inputPriority8->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString8, "")));
+	m_inputPriority8->setPosition(348, -46);
+	m_inputPriority8->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString8));
 	m_inputPriority8->setCommonFilter(CommonFilter::Any);
 	m_inputPriority8->setCallback([this](std::string const& str){
 
@@ -390,8 +393,8 @@ bool NotepadLayer::init(){
 	layout->addChild(m_inputPriority8);
 
     m_inputPriority9 = TextInput::create(50, "", "bigFont.fnt");
-	m_inputPriority9->setPosition(395, -86);
-	m_inputPriority9->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString9, "")));
+	m_inputPriority9->setPosition(348, -86);
+	m_inputPriority9->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString9));
 	m_inputPriority9->setCommonFilter(CommonFilter::Any);
 	m_inputPriority9->setCallback([this](std::string const& str){
 
@@ -401,8 +404,8 @@ bool NotepadLayer::init(){
 	layout->addChild(m_inputPriority9);
 
     m_inputPriority10 = TextInput::create(50, "", "bigFont.fnt");
-	m_inputPriority10->setPosition(395, -126);
-	m_inputPriority10->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString10, "")));
+	m_inputPriority10->setPosition(348, -126);
+	m_inputPriority10->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString10));
 	m_inputPriority10->setCommonFilter(CommonFilter::Any);
 	m_inputPriority10->setCallback([this](std::string const& str){
 
@@ -413,58 +416,103 @@ bool NotepadLayer::init(){
 
 	// Checkboxes
 
-	m_toggle1 = CCMenuItemToggler::createWithStandardSprites(this, menu_selector(NotepadLayer::saveCheckbox), 1);
+	m_toggle1 = CCMenuItemToggler::createWithStandardSprites(this, menu_selector(NotepadLayer::saveCheckbox1), 1);
 	layout->addChild(m_toggle1);
-	m_toggle1->setPosition(312, 234);
+	m_toggle1->setPosition(287, 234);
 	m_toggle1->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab1, false));
 
 	m_toggle2 = CCMenuItemToggler::createWithStandardSprites(this, menu_selector(NotepadLayer::saveCheckbox2), 1);
 	layout->addChild(m_toggle2);
-	m_toggle2->setPosition(312, 194);
+	m_toggle2->setPosition(287, 194);
 	m_toggle2->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab2, false));
 
 	m_toggle3 = CCMenuItemToggler::createWithStandardSprites(this, menu_selector(NotepadLayer::saveCheckbox3), 1);
 	layout->addChild(m_toggle3);
-	m_toggle3->setPosition(312, 154);
+	m_toggle3->setPosition(287, 154);
 	m_toggle3->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab3, false));
 
 	m_toggle4 = CCMenuItemToggler::createWithStandardSprites(this, menu_selector(NotepadLayer::saveCheckbox4), 1);
 	layout->addChild(m_toggle4);
-	m_toggle4->setPosition(312, 114);
+	m_toggle4->setPosition(287, 114);
 	m_toggle4->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab4, false));
 
 	m_toggle5 = CCMenuItemToggler::createWithStandardSprites(this, menu_selector(NotepadLayer::saveCheckbox5), 1);
 	layout->addChild(m_toggle5);
-	m_toggle5->setPosition(312, 74);
+	m_toggle5->setPosition(287, 74);
 	m_toggle5->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab5, false));
 
 	m_toggle6 = CCMenuItemToggler::createWithStandardSprites(this, menu_selector(NotepadLayer::saveCheckbox6), 1);
 	layout->addChild(m_toggle6);
-	m_toggle6->setPosition(312, 34);
+	m_toggle6->setPosition(287, 34);
 	m_toggle6->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab6, false));
 
 	m_toggle7 = CCMenuItemToggler::createWithStandardSprites(this, menu_selector(NotepadLayer::saveCheckbox7), 1);
 	layout->addChild(m_toggle7);
-	m_toggle7->setPosition(312, -6);
+	m_toggle7->setPosition(287, -6);
 	m_toggle7->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab7, false));
 
 	m_toggle8 = CCMenuItemToggler::createWithStandardSprites(this, menu_selector(NotepadLayer::saveCheckbox8), 1);
 	layout->addChild(m_toggle8);
-	m_toggle8->setPosition(312, -46);
+	m_toggle8->setPosition(287, -46);
 	m_toggle8->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab8, false));
 
 	m_toggle9 = CCMenuItemToggler::createWithStandardSprites(this, menu_selector(NotepadLayer::saveCheckbox9), 1);
 	layout->addChild(m_toggle9);
-	m_toggle9->setPosition(312, -86);
+	m_toggle9->setPosition(287, -86);
 	m_toggle9->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab9, false));
 
 	m_toggle10 = CCMenuItemToggler::createWithStandardSprites(this, menu_selector(NotepadLayer::saveCheckbox10), 1);
 	layout->addChild(m_toggle10);
-	m_toggle10->setPosition(312, -126);
+	m_toggle10->setPosition(287, -126);
 	m_toggle10->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab10, false));
 
+	// Clear Row Buttons
 
-	// Settings
+	auto onClearRowSpr = CCSprite::createWithSpriteFrameName("GJ_trashBtn_001.png");
+	onClearRowSpr->setScale(0.75f);
+
+	auto clearRowBtn1 = CCMenuItemSpriteExtra::create(onClearRowSpr, this, menu_selector(NotepadLayer::onClearRow1));
+	layout->addChild(clearRowBtn1);
+	clearRowBtn1->setPosition(404, 234);
+
+	auto clearRowBtn2 = CCMenuItemSpriteExtra::create(onClearRowSpr, this, menu_selector(NotepadLayer::onClearRow2));
+	layout->addChild(clearRowBtn2);
+	clearRowBtn2->setPosition(404, 194);
+
+	auto clearRowBtn3 = CCMenuItemSpriteExtra::create(onClearRowSpr, this, menu_selector(NotepadLayer::onClearRow3));
+	layout->addChild(clearRowBtn3);
+	clearRowBtn3->setPosition(404, 154);
+
+	auto clearRowBtn4 = CCMenuItemSpriteExtra::create(onClearRowSpr, this, menu_selector(NotepadLayer::onClearRow4));
+	layout->addChild(clearRowBtn4);
+	clearRowBtn4->setPosition(404, 114);
+
+	auto clearRowBtn5 = CCMenuItemSpriteExtra::create(onClearRowSpr, this, menu_selector(NotepadLayer::onClearRow5));
+	layout->addChild(clearRowBtn5);
+	clearRowBtn5->setPosition(404, 74);
+
+	auto clearRowBtn6 = CCMenuItemSpriteExtra::create(onClearRowSpr, this, menu_selector(NotepadLayer::onClearRow6));
+	layout->addChild(clearRowBtn6);
+	clearRowBtn6->setPosition(404, 34);
+
+	auto clearRowBtn7 = CCMenuItemSpriteExtra::create(onClearRowSpr, this, menu_selector(NotepadLayer::onClearRow7));
+	layout->addChild(clearRowBtn7);
+	clearRowBtn7->setPosition(404, -6);
+
+	auto clearRowBtn8 = CCMenuItemSpriteExtra::create(onClearRowSpr, this, menu_selector(NotepadLayer::onClearRow8));
+	layout->addChild(clearRowBtn8);
+	clearRowBtn8->setPosition(404, -46);
+
+	auto clearRowBtn9 = CCMenuItemSpriteExtra::create(onClearRowSpr, this, menu_selector(NotepadLayer::onClearRow9));
+	layout->addChild(clearRowBtn9);
+	clearRowBtn9->setPosition(404, -86);
+
+	auto clearRowBtn10 = CCMenuItemSpriteExtra::create(onClearRowSpr, this, menu_selector(NotepadLayer::onClearRow10));
+	layout->addChild(clearRowBtn10);
+	clearRowBtn10->setPosition(404, -126);
+
+
+	// Buttons
 
 	auto onSettingsSpr = CCSprite::createWithSpriteFrameName("GJ_optionsBtn_001.png");
 	onSettingsSpr->setScale(0.6f);
@@ -474,29 +522,20 @@ bool NotepadLayer::init(){
 	onSettingsBtn->setID("settings");
 	onSettingsBtn->setPosition(411, 254);
 
-	// pageMenu
+	auto onClearSpr = CCSprite::createWithSpriteFrameName("GJ_trashBtn_001.png");
+	onClearSpr->setScale(0.725f);
 
-	m_inputName = TextInput::create(100, "1/5", "bigFont.fnt");
-	m_inputName->setPosition(284.5f, 268);
-	m_inputName->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_savedValue, "")));
-	m_inputName->setCommonFilter(CommonFilter::Any);
-	m_inputName->setCallback([this](std::string const& str){
+	auto clearAllBtn = CCMenuItemSpriteExtra::create(onClearSpr, this, menu_selector(NotepadLayer::onClearAll));
+	m_buttonMenu->addChild(clearAllBtn);
+	clearAllBtn->setID("clear-all");
+	clearAllBtn->setPosition(375, 254);
 
-    (Mod::get()->setSavedValue<std::string>(m_savedValue, str));
+	// Level Name
 
-	});
-	pageMenu->addChild(m_inputName);
-
-	m_pageLeft = CCMenuItemSpriteExtra::create(CCSprite::createWithSpriteFrameName("edit_leftBtn_001.png"), this, menu_selector(NotepadLayer::onPage));
-	pageMenu->addChild(m_pageLeft);
-	m_pageLeft->setTag(4);
-	m_pageLeft->setPosition(221.0f, 268.0f);
-
-	auto pageRight = CCMenuItemSpriteExtra::create(CCSprite::createWithSpriteFrameName("edit_rightBtn_001.png"), this, menu_selector(NotepadLayer::onPage));
-	pageRight->setPosition(348.0f, 268.0f);
-    pageRight->setTag(1);
-    pageMenu->addChild(pageRight);
-
+	auto levelNameLabel = CCTextInputNode::create(250, 0, "", "goldFont.fnt");
+	levelNameLabel->setPosition(284.5f, 270);
+	levelNameLabel->setString(Mod::get()->getSavedValue<std::string>("editor-level-name"));
+	levelName->addChild(levelNameLabel);
 
 	// ScrollLayer
 
@@ -533,253 +572,191 @@ void NotepadLayer::onSettings(CCObject*){
 	this->removeFromParentAndCleanup(true);
 	geode::openSettingsPopup(Mod::get());
 }
+
+void NotepadLayer::onClearAll(CCObject*){
+
+	geode::createQuickPopup("Clear All", "Are you sure you want to <cr>clear</c> all rows?", "No", "Yes", [this](auto, bool yes) {
+        if (yes) {
+
+			m_input1->setString(Mod::get()->getSavedValue<std::string>(m_inputTab1, Mod::get()->setSavedValue<std::string>(m_inputTab1, "")));
+			m_input2->setString(Mod::get()->getSavedValue<std::string>(m_inputTab2, Mod::get()->setSavedValue<std::string>(m_inputTab2, "")));
+			m_input3->setString(Mod::get()->getSavedValue<std::string>(m_inputTab3, Mod::get()->setSavedValue<std::string>(m_inputTab3, "")));
+			m_input4->setString(Mod::get()->getSavedValue<std::string>(m_inputTab4, Mod::get()->setSavedValue<std::string>(m_inputTab4, "")));
+			m_input5->setString(Mod::get()->getSavedValue<std::string>(m_inputTab5, Mod::get()->setSavedValue<std::string>(m_inputTab5, "")));
+			m_input6->setString(Mod::get()->getSavedValue<std::string>(m_inputTab6, Mod::get()->setSavedValue<std::string>(m_inputTab6, "")));
+			m_input7->setString(Mod::get()->getSavedValue<std::string>(m_inputTab7, Mod::get()->setSavedValue<std::string>(m_inputTab7, "")));
+			m_input8->setString(Mod::get()->getSavedValue<std::string>(m_inputTab8, Mod::get()->setSavedValue<std::string>(m_inputTab8, "")));
+			m_input9->setString(Mod::get()->getSavedValue<std::string>(m_inputTab9, Mod::get()->setSavedValue<std::string>(m_inputTab9, "")));
+			m_input10->setString(Mod::get()->getSavedValue<std::string>(m_inputTab10, Mod::get()->setSavedValue<std::string>(m_inputTab10, "")));
+
+			m_inputPriority1->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString1, Mod::get()->setSavedValue<std::string>(m_inputPriorityString1, "")));
+			m_inputPriority2->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString2, Mod::get()->setSavedValue<std::string>(m_inputPriorityString2, "")));
+			m_inputPriority3->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString3, Mod::get()->setSavedValue<std::string>(m_inputPriorityString3, "")));
+			m_inputPriority4->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString4, Mod::get()->setSavedValue<std::string>(m_inputPriorityString4, "")));
+			m_inputPriority5->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString5, Mod::get()->setSavedValue<std::string>(m_inputPriorityString5, "")));
+			m_inputPriority6->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString6, Mod::get()->setSavedValue<std::string>(m_inputPriorityString6, "")));
+			m_inputPriority7->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString7, Mod::get()->setSavedValue<std::string>(m_inputPriorityString7, "")));
+			m_inputPriority8->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString8, Mod::get()->setSavedValue<std::string>(m_inputPriorityString8, "")));
+			m_inputPriority9->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString9, Mod::get()->setSavedValue<std::string>(m_inputPriorityString9, "")));
+			m_inputPriority10->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString10, Mod::get()->setSavedValue<std::string>(m_inputPriorityString10, "")));
+
+			m_toggle1->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab1, Mod::get()->setSavedValue<bool>(m_toggleTab1, false)));
+			m_toggle2->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab2, Mod::get()->setSavedValue<bool>(m_toggleTab2, false)));
+			m_toggle3->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab3, Mod::get()->setSavedValue<bool>(m_toggleTab3, false)));
+			m_toggle4->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab4, Mod::get()->setSavedValue<bool>(m_toggleTab4, false)));
+			m_toggle5->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab5, Mod::get()->setSavedValue<bool>(m_toggleTab5, false)));
+			m_toggle6->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab6, Mod::get()->setSavedValue<bool>(m_toggleTab6, false)));
+			m_toggle7->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab7, Mod::get()->setSavedValue<bool>(m_toggleTab7, false)));
+			m_toggle8->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab8, Mod::get()->setSavedValue<bool>(m_toggleTab8, false)));
+			m_toggle9->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab9, Mod::get()->setSavedValue<bool>(m_toggleTab9, false)));
+			m_toggle10->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab10, Mod::get()->setSavedValue<bool>(m_toggleTab10, false)));
+            
+        }
+    });
 	
-void NotepadLayer::onPage(CCObject* sender) {
-    m_page += sender->getTag();
-	m_scroll->scrollToTop();
-
-	if (m_page > 0){
-
-		m_inputTab1 = "input-1-tab-2";
-		m_inputTab2 = "input-2-tab-2";
-		m_inputTab3 = "input-3-tab-2";
-		m_inputTab4 = "input-4-tab-2";
-		m_inputTab5 = "input-5-tab-2";
-		m_inputTab6 = "input-6-tab-2";
-		m_inputTab7 = "input-7-tab-2";
-		m_inputTab8 = "input-8-tab-2";
-		m_inputTab9 = "input-9-tab-2";
-		m_inputTab10 = "input-10-tab-2";
-
-		m_inputPriorityString1 = "input-priority-one-tab-2";
-		m_inputPriorityString2 = "input-priority-two-tab-2";
-		m_inputPriorityString3 = "input-priority-three-tab-2";
-		m_inputPriorityString4 = "input-priority-four-tab-2";
-		m_inputPriorityString5 = "input-priority-five-tab-2";
-		m_inputPriorityString6 = "input-priority-six-tab-2";
-		m_inputPriorityString7 = "input-priority-seven-tab-2";
-		m_inputPriorityString8 = "input-priority-eight-tab-2";
-		m_inputPriorityString9 = "input-priority-nine-tab-2";
-		m_inputPriorityString10 = "input-priority-ten-tab-2";
-
-		m_toggleTab1 = "checkbox-tab-2";
-		m_toggleTab2 = "checkbox-2-tab-2";
-		m_toggleTab3 = "checkbox-3-tab-2";
-		m_toggleTab4 = "checkbox-4-tab-2";
-		m_toggleTab5 = "checkbox-5-tab-2";
-		m_toggleTab6 = "checkbox-6-tab-2";
-		m_toggleTab7 = "checkbox-7-tab-2";
-		m_toggleTab8 = "checkbox-8-tab-2";
-		m_toggleTab9 = "checkbox-9-tab-2";
-		m_toggleTab10 = "checkbox-10-tab-2";
-	
-		m_savedValue = "input-name-1-tab-2";
-		m_inputName->setPlaceholder("2/5");
-	 	m_pageLeft->setTag(-2);	
-				
-	}
-	if (m_page > 1){
-
-		m_inputTab1 = "input-1-tab-3";
-		m_inputTab2 = "input-2-tab-3";
-		m_inputTab3 = "input-3-tab-3";
-		m_inputTab4 = "input-4-tab-3";
-		m_inputTab5 = "input-5-tab-3";
-		m_inputTab6 = "input-6-tab-3";
-		m_inputTab7 = "input-7-tab-3";
-		m_inputTab8 = "input-8-tab-3";
-		m_inputTab9 = "input-9-tab-3";
-		m_inputTab10 = "input-10-tab-3";
-
-		m_inputPriorityString1 = "input-priority-one-tab-3";
-		m_inputPriorityString2 = "input-priority-two-tab-3";
-		m_inputPriorityString3 = "input-priority-three-tab-3";
-		m_inputPriorityString4 = "input-priority-four-tab-3";
-		m_inputPriorityString5 = "input-priority-five-tab-3";
-		m_inputPriorityString6 = "input-priority-six-tab-3";
-		m_inputPriorityString7 = "input-priority-seven-tab-3";
-		m_inputPriorityString8 = "input-priority-eight-tab-3";
-		m_inputPriorityString9 = "input-priority-nine-tab-3";
-		m_inputPriorityString10 = "input-priority-ten-tab-3";
-
-		m_toggleTab1 = "checkbox-tab-3";
-		m_toggleTab2 = "checkbox-2-tab-3";
-		m_toggleTab3 = "checkbox-3-tab-3";
-		m_toggleTab4 = "checkbox-4-tab-3";
-		m_toggleTab5 = "checkbox-5-tab-3";
-		m_toggleTab6 = "checkbox-6-tab-3";
-		m_toggleTab7 = "checkbox-7-tab-3";
-		m_toggleTab8 = "checkbox-8-tab-3";
-		m_toggleTab9 = "checkbox-9-tab-3";
-		m_toggleTab10 = "checkbox-10-tab-3";
-
-		m_savedValue = "input-name-1-tab-3";
-		m_inputName->setPlaceholder("3/5");
-		m_pageLeft->setTag(-1);
-				
-	}
-	if (m_page > 2){
-
-		m_inputTab1 = "input-1-tab-4";
-		m_inputTab2 = "input-2-tab-4";
-		m_inputTab3 = "input-3-tab-4";
-		m_inputTab4 = "input-4-tab-4";
-		m_inputTab5 = "input-5-tab-4";
-		m_inputTab6 = "input-6-tab-4";
-		m_inputTab7 = "input-7-tab-4";
-		m_inputTab8 = "input-8-tab-4";
-		m_inputTab9 = "input-9-tab-4";
-		m_inputTab10 = "input-10-tab-4";
-
-		m_inputPriorityString1 = "input-priority-one-tab-4";
-		m_inputPriorityString2 = "input-priority-two-tab-4";
-		m_inputPriorityString3 = "input-priority-three-tab-4";
-		m_inputPriorityString4 = "input-priority-four-tab-4";
-		m_inputPriorityString5 = "input-priority-five-tab-4";
-		m_inputPriorityString6 = "input-priority-six-tab-4";
-		m_inputPriorityString7 = "input-priority-seven-tab-4";
-		m_inputPriorityString8 = "input-priority-eight-tab-4";
-		m_inputPriorityString9 = "input-priority-nine-tab-4";
-		m_inputPriorityString10 = "input-priority-ten-tab-4";
-		
-		m_toggleTab1 = "checkbox-tab-4";
-		m_toggleTab2 = "checkbox-2-tab-4";
-		m_toggleTab3 = "checkbox-3-tab-4";
-		m_toggleTab4 = "checkbox-4-tab-4";
-		m_toggleTab5 = "checkbox-5-tab-4";
-		m_toggleTab6 = "checkbox-6-tab-4";
-		m_toggleTab7 = "checkbox-7-tab-4";
-		m_toggleTab8 = "checkbox-8-tab-4";
-		m_toggleTab9 = "checkbox-9-tab-4";
-		m_toggleTab10 = "checkbox-10-tab-4";
-
-		m_savedValue = "input-name-1-tab-4";
-		m_inputName->setPlaceholder("4/5");
-
-	}
-	if (m_page > 3){
-
-		m_inputTab1 = "input-1-tab-5";
-		m_inputTab2 = "input-2-tab-5";
-		m_inputTab3 = "input-3-tab-5";
-		m_inputTab4 = "input-4-tab-5";
-		m_inputTab5 = "input-5-tab-5";
-		m_inputTab6 = "input-6-tab-5";
-		m_inputTab7 = "input-7-tab-5";
-		m_inputTab8 = "input-8-tab-5";
-		m_inputTab9 = "input-9-tab-5";
-		m_inputTab10 = "input-10-tab-5";
-
-		m_inputPriorityString1 = "input-priority-one-tab-5";
-		m_inputPriorityString2 = "input-priority-two-tab-5";
-		m_inputPriorityString3 = "input-priority-three-tab-5";
-		m_inputPriorityString4 = "input-priority-four-tab-5";
-		m_inputPriorityString5 = "input-priority-five-tab-5";
-		m_inputPriorityString6 = "input-priority-six-tab-5";
-		m_inputPriorityString7 = "input-priority-seven-tab-5";
-		m_inputPriorityString8 = "input-priority-eight-tab-5";
-		m_inputPriorityString9 = "input-priority-nine-tab-5";
-		m_inputPriorityString10 = "input-priority-ten-tab-5";
-
-		m_toggleTab1 = "checkbox-tab-5";
-		m_toggleTab2 = "checkbox-2-tab-5";
-		m_toggleTab3 = "checkbox-3-tab-5";
-		m_toggleTab4 = "checkbox-4-tab-5";
-		m_toggleTab5 = "checkbox-5-tab-5";
-		m_toggleTab6 = "checkbox-6-tab-5";
-		m_toggleTab7 = "checkbox-7-tab-5";
-		m_toggleTab8 = "checkbox-8-tab-5";
-		m_toggleTab9 = "checkbox-9-tab-5";
-		m_toggleTab10 = "checkbox-10-tab-5";
-
-		m_savedValue = "input-name-1-tab-5";
-		m_inputName->setPlaceholder("5/5");
-		m_pageLeft->setTag(-1);
-				
-	}
-	if (m_page > 4){
-	
-		m_inputTab1 = "input-1-tab-1";
-		m_inputTab2 = "input-2-tab-1";
-		m_inputTab3 = "input-3-tab-1";
-		m_inputTab4 = "input-4-tab-1";
-		m_inputTab5 = "input-5-tab-1";
-		m_inputTab6 = "input-6-tab-1";
-		m_inputTab7 = "input-7-tab-1";
-		m_inputTab8 = "input-8-tab-1";
-		m_inputTab9 = "input-9-tab-1";
-		m_inputTab10 = "input-10-tab-1";
-
-		m_inputPriorityString1 = "input-priority-one-tab-1";
-		m_inputPriorityString2 = "input-priority-two-tab-1";
-		m_inputPriorityString3 = "input-priority-three-tab-1";
-		m_inputPriorityString4 = "input-priority-four-tab-1";
-		m_inputPriorityString5 = "input-priority-five-tab-1";
-		m_inputPriorityString6 = "input-priority-six-tab-1";
-		m_inputPriorityString7 = "input-priority-seven-tab-1";
-		m_inputPriorityString8 = "input-priority-eight-tab-1";
-		m_inputPriorityString9 = "input-priority-nine-tab-1";
-		m_inputPriorityString10 = "input-priority-ten-tab-1";
-
-		m_toggleTab1 = "checkbox-tab-1";
-		m_toggleTab2 = "checkbox-2-tab-1";
-		m_toggleTab3 = "checkbox-3-tab-1";
-		m_toggleTab4 = "checkbox-4-tab-1";
-		m_toggleTab5 = "checkbox-5-tab-1";
-		m_toggleTab6 = "checkbox-6-tab-1";
-		m_toggleTab7 = "checkbox-7-tab-1";
-		m_toggleTab8 = "checkbox-8-tab-1";
-		m_toggleTab9 = "checkbox-9-tab-1";
-		m_toggleTab10 = "checkbox-10-tab-1";
-
-		m_savedValue = "input-name-1-tab-1";
-		m_inputName->setPlaceholder("1/5");
-		m_pageLeft->setTag(4);
-		
-		m_page = 0;
-	}
-
-	// TextInputs
-
-	m_inputName->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_savedValue, "")));
-
-	m_input1->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab1, "")));
-	m_input2->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab2, "")));
-	m_input3->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab3, "")));
-	m_input4->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab4, "")));
-	m_input5->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab5, "")));
-	m_input6->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab6, "")));
-	m_input7->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab7, "")));
-	m_input8->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab8, "")));
-	m_input9->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab9, "")));
-	m_input10->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputTab10, "")));
-
-	m_inputPriority1->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString1, "")));
-	m_inputPriority2->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString2, "")));
-	m_inputPriority3->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString3, "")));
-	m_inputPriority4->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString4, "")));
-	m_inputPriority5->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString5, "")));
-	m_inputPriority6->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString6, "")));
-	m_inputPriority7->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString7, "")));
-	m_inputPriority8->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString8, "")));
-	m_inputPriority9->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString9, "")));
-	m_inputPriority10->setString(fmt::format("{}", Mod::get()->getSavedValue<std::string>(m_inputPriorityString10, "")));
-
-	// Checkboxes
-
-	m_toggle1->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab1, false));
-	m_toggle2->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab2, false));
-	m_toggle3->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab3, false));
-	m_toggle4->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab4, false));
-	m_toggle5->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab5, false));
-	m_toggle6->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab6, false));
-	m_toggle7->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab7, false));
-	m_toggle8->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab8, false));
-	m_toggle9->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab9, false));
-	m_toggle10->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab10, false));
 }
 
-void NotepadLayer::saveCheckbox(CCObject*){
+void NotepadLayer::onClearRow1(CCObject*){
+
+	geode::createQuickPopup("Clear Row", "Are you sure you want to <cr>clear</c> this row?", "No", "Yes", [this](auto, bool yes) {
+        if (yes) {
+
+			m_input1->setString(Mod::get()->getSavedValue<std::string>(m_inputTab1, Mod::get()->setSavedValue<std::string>(m_inputTab1, "")));
+			m_inputPriority1->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString1, Mod::get()->setSavedValue<std::string>(m_inputPriorityString1, "")));
+			m_toggle1->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab1, Mod::get()->setSavedValue<bool>(m_toggleTab1, false)));
+            
+        }
+    });
+	
+}
+
+void NotepadLayer::onClearRow2(CCObject*){
+
+	geode::createQuickPopup("Clear Row", "Are you sure you want to <cr>clear</c> this row?", "No", "Yes", [this](auto, bool yes) {
+        if (yes) {
+
+			m_input2->setString(Mod::get()->getSavedValue<std::string>(m_inputTab2, Mod::get()->setSavedValue<std::string>(m_inputTab2, "")));
+			m_inputPriority2->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString2, Mod::get()->setSavedValue<std::string>(m_inputPriorityString2, "")));
+			m_toggle2->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab2, Mod::get()->setSavedValue<bool>(m_toggleTab2, false)));
+            
+        }
+    });
+	
+}
+
+void NotepadLayer::onClearRow3(CCObject*){
+
+	geode::createQuickPopup("Clear Row", "Are you sure you want to <cr>clear</c> this row?", "No", "Yes", [this](auto, bool yes) {
+        if (yes) {
+
+			m_input3->setString(Mod::get()->getSavedValue<std::string>(m_inputTab3, Mod::get()->setSavedValue<std::string>(m_inputTab3, "")));
+			m_inputPriority3->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString3, Mod::get()->setSavedValue<std::string>(m_inputPriorityString3, "")));
+			m_toggle3->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab3, Mod::get()->setSavedValue<bool>(m_toggleTab3, false)));
+            
+        }
+    });
+	
+}
+
+void NotepadLayer::onClearRow4(CCObject*){
+
+	geode::createQuickPopup("Clear Row", "Are you sure you want to <cr>clear</c> this row?", "No", "Yes", [this](auto, bool yes) {
+        if (yes) {
+
+			m_input4->setString(Mod::get()->getSavedValue<std::string>(m_inputTab4, Mod::get()->setSavedValue<std::string>(m_inputTab4, "")));
+			m_inputPriority4->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString4, Mod::get()->setSavedValue<std::string>(m_inputPriorityString4, "")));
+			m_toggle4->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab4, Mod::get()->setSavedValue<bool>(m_toggleTab4, false)));
+            
+        }
+    });
+	
+}
+
+void NotepadLayer::onClearRow5(CCObject*){
+
+	geode::createQuickPopup("Clear Row", "Are you sure you want to <cr>clear</c> this row?", "No", "Yes", [this](auto, bool yes) {
+        if (yes) {
+
+			m_input5->setString(Mod::get()->getSavedValue<std::string>(m_inputTab5, Mod::get()->setSavedValue<std::string>(m_inputTab5, "")));
+			m_inputPriority5->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString5, Mod::get()->setSavedValue<std::string>(m_inputPriorityString5, "")));
+			m_toggle5->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab5, Mod::get()->setSavedValue<bool>(m_toggleTab5, false)));
+            
+        }
+    });
+	
+}
+
+void NotepadLayer::onClearRow6(CCObject*){
+
+	geode::createQuickPopup("Clear Row", "Are you sure you want to <cr>clear</c> this row?", "No", "Yes", [this](auto, bool yes) {
+        if (yes) {
+
+			m_input6->setString(Mod::get()->getSavedValue<std::string>(m_inputTab6, Mod::get()->setSavedValue<std::string>(m_inputTab6, "")));
+			m_inputPriority6->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString6, Mod::get()->setSavedValue<std::string>(m_inputPriorityString6, "")));
+			m_toggle6->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab6, Mod::get()->setSavedValue<bool>(m_toggleTab6, false)));
+            
+        }
+    });
+	
+}
+
+void NotepadLayer::onClearRow7(CCObject*){
+
+	geode::createQuickPopup("Clear Row", "Are you sure you want to <cr>clear</c> this row?", "No", "Yes", [this](auto, bool yes) {
+        if (yes) {
+
+			m_input7->setString(Mod::get()->getSavedValue<std::string>(m_inputTab7, Mod::get()->setSavedValue<std::string>(m_inputTab7, "")));
+			m_inputPriority7->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString7, Mod::get()->setSavedValue<std::string>(m_inputPriorityString7, "")));
+			m_toggle7->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab7, Mod::get()->setSavedValue<bool>(m_toggleTab7, false)));
+            
+        }
+    });
+	
+}
+
+void NotepadLayer::onClearRow8(CCObject*){
+
+	geode::createQuickPopup("Clear Row", "Are you sure you want to <cr>clear</c> this row?", "No", "Yes", [this](auto, bool yes) {
+        if (yes) {
+
+			m_input8->setString(Mod::get()->getSavedValue<std::string>(m_inputTab8, Mod::get()->setSavedValue<std::string>(m_inputTab8, "")));
+			m_inputPriority8->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString8, Mod::get()->setSavedValue<std::string>(m_inputPriorityString8, "")));
+			m_toggle8->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab8, Mod::get()->setSavedValue<bool>(m_toggleTab8, false)));
+            
+        }
+    });
+	
+}
+
+void NotepadLayer::onClearRow9(CCObject*){
+
+	geode::createQuickPopup("Clear Row", "Are you sure you want to <cr>clear</c> this row?", "No", "Yes", [this](auto, bool yes) {
+        if (yes) {
+
+			m_input9->setString(Mod::get()->getSavedValue<std::string>(m_inputTab9, Mod::get()->setSavedValue<std::string>(m_inputTab9, "")));
+			m_inputPriority9->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString9, Mod::get()->setSavedValue<std::string>(m_inputPriorityString9, "")));
+			m_toggle9->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab9, Mod::get()->setSavedValue<bool>(m_toggleTab9, false)));
+            
+        }
+    });
+	
+}
+
+void NotepadLayer::onClearRow10(CCObject*){
+
+	geode::createQuickPopup("Clear Row", "Are you sure you want to <cr>clear</c> this row?", "No", "Yes", [this](auto, bool yes) {
+        if (yes) {
+
+			m_input10->setString(Mod::get()->getSavedValue<std::string>(m_inputTab10, Mod::get()->setSavedValue<std::string>(m_inputTab10, "")));
+			m_inputPriority10->setString(Mod::get()->getSavedValue<std::string>(m_inputPriorityString10, Mod::get()->setSavedValue<std::string>(m_inputPriorityString10, "")));
+			m_toggle10->toggle(Mod::get()->getSavedValue<bool>(m_toggleTab10, Mod::get()->setSavedValue<bool>(m_toggleTab10, false)));
+            
+        }
+    });
+	
+}
+
+void NotepadLayer::saveCheckbox1(CCObject*){
 	Mod::get()->setSavedValue<bool>(m_toggleTab1, !Mod::get()->getSavedValue<bool>(m_toggleTab1, false));
 }