+++ src/NotepadLayer.hpp
@@ -2,47 +2,43 @@
 
 using namespace geode::prelude;
 
-class NotepadLayer: public Popup{
+class NotepadLayer: public Popup, public FLAlertLayerProtocol{
 protected:
 
-const char* m_savedValue = "input-name-1-tab-1";
+std::string m_levelID = Mod::get()->getSavedValue<std::string>("editor-level-id");
 
-const char* m_inputTab1 = "input-1-tab-1";
-const char* m_inputTab2 = "input-2-tab-1";
-const char* m_inputTab3 = "input-3-tab-1";
-const char* m_inputTab4 = "input-4-tab-1";
-const char* m_inputTab5 = "input-5-tab-1";
-const char* m_inputTab6 = "input-6-tab-1";
-const char* m_inputTab7 = "input-7-tab-1";
-const char* m_inputTab8 = "input-8-tab-1";
-const char* m_inputTab9 = "input-9-tab-1";
-const char* m_inputTab10 = "input-10-tab-1";
+std::string m_inputTab1 = m_levelID + "-input-1";
+std::string m_inputTab2 = m_levelID + "-input-2";
+std::string m_inputTab3 = m_levelID + "-input-3";
+std::string m_inputTab4 = m_levelID + "-input-4";
+std::string m_inputTab5 = m_levelID + "-input-5";
+std::string m_inputTab6 = m_levelID + "-input-6";
+std::string m_inputTab7 = m_levelID + "-input-7";
+std::string m_inputTab8 = m_levelID + "-input-8";
+std::string m_inputTab9 = m_levelID + "-input-9";
+std::string m_inputTab10 = m_levelID + "-input-10";
 
-const char* m_inputPriorityString1 = "input-priority-one-tab-1";
-const char* m_inputPriorityString2 = "input-priority-two-tab-1";
-const char* m_inputPriorityString3 = "input-priority-three-tab-1";
-const char* m_inputPriorityString4 = "input-priority-four-tab-1";
-const char* m_inputPriorityString5 = "input-priority-five-tab-1";
-const char* m_inputPriorityString6 = "input-priority-six-tab-1";
-const char* m_inputPriorityString7 = "input-priority-seven-tab-1";
-const char* m_inputPriorityString8 = "input-priority-eight-tab-1";
-const char* m_inputPriorityString9 = "input-priority-nine-tab-1";
-const char* m_inputPriorityString10 = "input-priority-ten-tab-1";
+std::string m_inputPriorityString1 = m_levelID + "-input-priority-1";
+std::string m_inputPriorityString2 = m_levelID + "-input-priority-2";
+std::string m_inputPriorityString3 = m_levelID + "-input-priority-3";
+std::string m_inputPriorityString4 = m_levelID + "-input-priority-4";
+std::string m_inputPriorityString5 = m_levelID + "-input-priority-5";
+std::string m_inputPriorityString6 = m_levelID + "-input-priority-6";
+std::string m_inputPriorityString7 = m_levelID + "-input-priority-7";
+std::string m_inputPriorityString8 = m_levelID + "-input-priority-8";
+std::string m_inputPriorityString9 = m_levelID + "-input-priority-9";
+std::string m_inputPriorityString10 = m_levelID + "-input-priority-10";
 
-const char* m_toggleTab1 = "checkbox-tab-1";
-const char* m_toggleTab2 = "checkbox-2-tab-1";
-const char* m_toggleTab3 = "checkbox-3-tab-1";
-const char* m_toggleTab4 = "checkbox-4-tab-1";
-const char* m_toggleTab5 = "checkbox-5-tab-1";
-const char* m_toggleTab6 = "checkbox-6-tab-1";
-const char* m_toggleTab7 = "checkbox-7-tab-1";
-const char* m_toggleTab8 = "checkbox-8-tab-1";
-const char* m_toggleTab9 = "checkbox-9-tab-1";
-const char* m_toggleTab10 = "checkbox-10-tab-1";
-
-size_t m_page;
-
-CCMenuItemSpriteExtra* m_pageLeft;
+std::string m_toggleTab1 = m_levelID + "-checkbox-1";
+std::string m_toggleTab2 = m_levelID + "-checkbox-2";
+std::string m_toggleTab3 = m_levelID + "-checkbox-3";
+std::string m_toggleTab4 = m_levelID + "-checkbox-4";
+std::string m_toggleTab5 = m_levelID + "-checkbox-5";
+std::string m_toggleTab6 = m_levelID + "-checkbox-6";
+std::string m_toggleTab7 = m_levelID + "-checkbox-7";
+std::string m_toggleTab8 = m_levelID + "-checkbox-8";
+std::string m_toggleTab9 = m_levelID + "-checkbox-9";
+std::string m_toggleTab10 = m_levelID + "-checkbox-10";
 
 CCMenuItemToggler* m_toggle1;
 CCMenuItemToggler* m_toggle2;
@@ -55,10 +51,7 @@ CCMenuItemToggler* m_toggle8;
 CCMenuItemToggler* m_toggle9;
 CCMenuItemToggler* m_toggle10;
 
-
 ScrollLayer* m_scroll;
-
-TextInput* m_inputName;
 
 TextInput* m_input1;
 TextInput* m_input2;
@@ -86,9 +79,20 @@ TextInput* m_inputPriority10;
 bool init();
 
 void onSettings(CCObject*);
-void onPage(CCObject* sender);
+void onClearAll(CCObject*);
 
-void saveCheckbox(CCObject*);
+void onClearRow1(CCObject*);
+void onClearRow2(CCObject*);
+void onClearRow3(CCObject*);
+void onClearRow4(CCObject*);
+void onClearRow5(CCObject*);
+void onClearRow6(CCObject*);
+void onClearRow7(CCObject*);
+void onClearRow8(CCObject*);
+void onClearRow9(CCObject*);
+void onClearRow10(CCObject*);
+
+void saveCheckbox1(CCObject*);
 void saveCheckbox2(CCObject*);
 void saveCheckbox3(CCObject*);
 void saveCheckbox4(CCObject*);